Checking if it is a Mountain Array
The arr is a mountain array if and only if:
arr.length > 3
The array must increasing to a peak value and must decreasing after the peak value like a mountain.
Eg: arr={0,1,3,4,6,8,5,3,2} ,here 8 is the peak value, so before the peak value the array is increasing and after the peak value the array is decreasing. So, the given array is mountainn array.